+++ erdjs/src/smartcontracts/codec/constants.ts
@@ -1,0 +1,1 @@
+export const SizeOfU32 = 4;