+++ erdjs/src/smartcontracts/index.ts
@@ -3,11 +3,11 @@
  * @module smartcontracts
  */
 
-export * from "./argument";
 export * from "./code";
 export * from "./codeMetadata";
 export * from "./function";
 export * from "./interface";
 export * from "./query";
+export * from "./returnCode";
 export * from "./smartContract";
 export * from "./transactionPayloadBuilders";