+++ erdjs/src/smartcontracts/typesystem/generic.ts
@@ -1,28 +1,53 @@
 import { guardValueIsSet } from "../../utils";
-import { isTyped, Type, TypedValue } from "./types";
+import { Type, TypedValue, NullType, TypePlaceholder } from "./types";
 
-export class OptionalType extends Type {
-    static One = new OptionalType();
+export class OptionType extends Type {
+    constructor(typeParameter: Type) {
+        super("Option", [typeParameter]);
+    }
 
-    private constructor() {
-        super("Optional");
+    isAssignableFrom(type: Type): boolean {
+        if (!(type instanceof OptionType)) {
+            return false;
+        }
+
+        let invariantTypeParameters = this.getFirstTypeParameter().equals(type.getFirstTypeParameter());
+        let fakeCovarianceToNull = type.getFirstTypeParameter() instanceof NullType;
+        return invariantTypeParameters || fakeCovarianceToNull;
     }
 }
 
-export class VectorType extends Type {
-    static One = new VectorType();
-
-    private constructor() {
-        super("Vector");
+export class ListType extends Type {
+    constructor(typeParameter: Type) {
+        super("List", [typeParameter]);
     }
 }
 
-export class OptionalValue extends TypedValue {
+export class OptionValue extends TypedValue {
     private readonly value: TypedValue | null;
 
-    constructor(value: TypedValue | null = null) {
-        super();
+    constructor(type: OptionType, value: TypedValue | null = null) {
+        super(type);
+
+        // TODO: assert value is of type type.getFirstTypeParameter()
+
         this.value = value;
+    }
+
+    /**
+     * Creates an OptionValue, as a missing option argument.
+     */
+    static newMissing(): OptionValue {
+        let type = new OptionType(new NullType());
+        return new OptionValue(type);
+    }
+
+    /**
+     * Creates an OptionValue, as a provided option argument.
+     */
+    static newProvided(typedValue: TypedValue): OptionValue {
+        let type = new OptionType(typedValue.getType());
+        return new OptionValue(type, typedValue);
     }
 
     isSet(): boolean {
@@ -38,22 +63,36 @@ export class OptionalValue extends TypedValue {
         return this.value ? this.value.valueOf() : null;
     }
 
-    equals(other: OptionalValue): boolean {
+    equals(other: OptionValue): boolean {
         return this.value?.equals(other.value) || false;
-    }
-
-    getType(): Type {
-        return OptionalType.One;
     }
 }
 
-// TODO: make generic (in TypeScript's sense) or check homogenity.
-export class Vector extends TypedValue {
+// TODO: Rename to ListValue, for consistency (though the term is slighly unfortunate).
+// Question for review: or not?
+export class List extends TypedValue {
     private readonly items: TypedValue[];
 
-    constructor(items: TypedValue[]) {
-        super();
+    /**
+     * 
+     * @param type the type of this TypedValue (an instance of ListType), not the type parameter of the ListType
+     * @param items the items, having the type type.getFirstTypeParameter()
+     */
+    constructor(type: ListType, items: TypedValue[]) {
+        super(type);
+
+        // TODO: assert items are of type type.getFirstTypeParameter()
+
         this.items = items;
+    }
+
+    static fromItems(items: TypedValue[]): List {
+        if (items.length == 0) {
+            return new List(new TypePlaceholder(), []);
+        }
+        
+        let typeParameter = items[0].getType();
+        return new List(typeParameter, items);
     }
 
     getLength(): number {
@@ -68,7 +107,7 @@ export class Vector extends TypedValue {
         return this.items.map(item => item.valueOf());
     }
 
-    equals(other: Vector): boolean {
+    equals(other: List): boolean {
         if (this.getLength() != other.getLength()) {
             return false;
         }
@@ -83,9 +122,5 @@ export class Vector extends TypedValue {
         }
 
         return true;
-    }
-
-    getType(): Type {
-        return VectorType.One;
     }
 }